"Fixed" the bug with KeepStockedBehaviour not adding initial items
No idea what is causing this. For some reason some behaviours are not
having initialize() called, but I have no idea why that is. I counted
using lldb and the number of entities with KeepStockedBehaviour is
about 24000, but there are only around 20000 calls to
KeepStockedBehaviour::initialize(). No idea why this is.

For now I will just manually add the initial item in the constructor
of the entity and hope that this doesn't also fail. Worrying bug!